requirements.txt
mkdocs
mkdocs-material
mkdocs-with-pdf
mkdocs-glightbox
mkdocs-git-revision-date-localized-plugin
mkdocs-open-in-new-tab
mkdocs-render-swagger-plugin
PyYAML